Adapting PushButton for multicomponent
Adapted PushButtonHelper to have support for multiple buttons in a
single circuit. Changed MultiLEDController to be MultiCompController and
will run all multi component tests through there now

PushButtonHelper no longer has a PostConstruct annotation. Due to
limitations with the annotation the method was not being executed
properly and was causing a bug, the method is now instead called at the
end of the constructor