Completes Issue #71 (#72)
* adding a new component adoc file

* started outlining the things needed for a new component and controller example

* more notes on new components

The document is just notes at this point. It will need to be cleaned up
once we learn more about the framework.

* outline of how to add new component

The outline is more clean than before but still lacks precise
information. Will need to double check if all the appropriate steps are
there and get the specific information.

* completed outlining how to add a new componenet in the documentation